// src/app/Modules/professor/Models/Work.model.ts
export class Work{
    id: number;
    subject: string;
    lastUpdate: Date;
    question: string;
    correction: string;
    explanation: string;
    firstHint: string;
    secondHint:string;
    categoryId: number;
    schoolYear: number;
    trimester: number;
    schoolYearCategoryId: number;
}